Refactor: Delete note/folder use cases can now throw Note/FolderNoteFoundError

+++ application/use_cases/folder/DeleteFolder.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 
+from domain.errors.FolderNotFoundError import FolderNotFoundError
 from domain.interfaces.IFolderRepository import IFolderRepository
 
 
@@ -11,4 +12,7 @@ class DeleteFolder:
         self._folder_repository = folder_repository
 
     def execute(self, folder_uuid: UUID) -> None:
+        folder = self._folder_repository.get_by_uuid(folder_uuid)
+        if folder is None:
+            raise FolderNotFoundError(f"Folder with uuid {folder_uuid} not found")
         self._folder_repository.delete(folder_uuid)
