+++ application/use_cases/note/DeleteNote.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 
+from domain.errors.NoteNotFoundError import NoteNotFoundError
 from domain.interfaces.INoteRepository import INoteRepository
 
 
@@ -8,4 +9,7 @@ class DeleteNote:
         self._note_repository = note_repository
 
     def execute(self, note_uuid: UUID) -> None:
+        note = self._note_repository.get_by_uuid(note_uuid)
+        if note is None:
+            raise NoteNotFoundError(f"Note with uuid {note_uuid} not found")
         self._note_repository.delete(note_uuid)
